pyproject.toml
[tool.poetry]
name = "discordduckbot-1"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mcstatus = "^3.1"
fuzzywuzzy = "^0.18.0"
flask = "^1.1"
"discord.py" = "^1.3.4"
dill = "^0.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
